dataflow-pipelines/src/bronze/requirements.txt
apache-beam[gcp]==2.56.0
google-cloud-bigquery
google-cloud-storage
